initia commit
For now there is not much. Only the stdin json reading system was
implemented. It works asynchronously with the usage of threads, so the
esc can compute further without it getting blocked.

Also the first thoughts about the architecture are applied here. The
most notable thing is, that square is divided into frontend and backend,
they are communicating though stdin and stdout.