InputHandler => IpcHandler
After a bit of thinking, the InputHandler, which from now on, i will
refer to as IpcHandler, should not only receive input, but also process
it. I am not quite sure if it is right to implement it as a system in
the ECS, but i will try it for now.