join IPC method implemented
IpcHandler is now ready to accept requests from other programs. Although
it is not much, the structure is now build and can be expanded without
much effort.

Since i was searching quite some while for best practices in the field
of IPC, or rather RPC (but through STDIN, not network) and
frameworks/libraries, that help to build these kind of application, i
was surprised that i have did not stumble upon anything really usefull.
tarpc looked very promising, but it was too network-oriented. Maybe i am
also really looking at things from the false perspective.

For now, i will continue to write in this style, even though contains a
lot of redundancy.